requirements.txt
Flask==1.1.1
neo4j-driver==4.2.0
gunicorn==20.0.4
pandas==1.0.1
selenium==3.141.0
Flask-Cors==3.0.9
Flask-Jsonpify==1.5.0
beautifulsoup4==4.8.2